Legends Z-A Support - v1.2.0 brings a lot of new changes!   - Pokémon Legends Z-A Support   - Automatic Bot Recovery System   - Advanced Per-Bot Logging System
## Contributors Welcome!
We are needing experienced members of the automations community to help us test and introduce various new features and tweaks to help advance PokeBot for the public.

Known Issues for PLZA:
1.  The Trader1MyStatusPointer is about 90% correct most of the time, the other 10% it loads empty data.  This is used for applying trade partner details (Auto OT).
2.  Clone/Dump commands are commented out and non-working.
Dump issues:  The pointer used views the trade partners pokemon without them having to select offer, so the data changes every time the user selects a new slot in their box.  The file is not written correctly, causing stats to be wrong when genned in the game via trade bot.
Clone issues: Clone works afaik, but is commented out due to suspicions that it might be barred later due to current game exploits.

## Notes about PLZA
1.  Due to Night/Daytime shift, users must be inside a building like Hotel Z to trade to prevent disruptions.
2.  Make sure you're on Box 1 Slot 1 before starting a trade.
3.  Make sure text speed is set to  Fast in the settings.